code review remarks resolved

+++ src/main/java/dev/eventmanager/entity/EventEntity.java
@@ -5,8 +5,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -59,7 +57,7 @@ public class EventEntity {
     private Integer duration;
 
     @Column(nullable = false)
-    private Integer locationId;
+    private Long locationId;
 
     @Column(nullable = false)
     private EventStatus status;
